warning fix

+++ SpaceGame/GameObjects/RectRenderer.cpp
@@ -190,7 +190,7 @@ void RectRenderer::Render(DxDevice *dxDev) {
             //GeometryFactory::CreateRectangleFilled(1.0f, 1.0f, 0.0f, pos, adjPrev, adjNext, aaDir, indices2);
             GeometryFactory::CreateRectangleFilled(1.0f, 1.0f, std::fabsf(std::sin(angle2)), pos, adjPrev, adjNext, aaDir, indices2);
 
-            this->indexCount = indices2.size();
+            this->indexCount = static_cast<uint32_t>(indices2.size());
             /*GeometryFactory::CreateRectangle(1.0f, 1.0f, 0.01f,
             DataBuffer<DirectX::XMFLOAT2>(vertices, &vertices->pos),
             DataBuffer<DirectX::XMFLOAT2>(vertices, &vertices->aaVec),
@@ -225,7 +225,7 @@ void RectRenderer::Render(DxDevice *dxDev) {
             D3D11_BUFFER_DESC bufDesc;
             D3D11_SUBRESOURCE_DATA initData;
 
-            bufDesc.ByteWidth = vertices.size() * sizeof(ColorVertex2D);
+            bufDesc.ByteWidth = static_cast<UINT>(vertices.size() * sizeof(ColorVertex2D));
             bufDesc.Usage = D3D11_USAGE_DEFAULT;
             bufDesc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
             bufDesc.CPUAccessFlags = 0;
@@ -238,7 +238,7 @@ void RectRenderer::Render(DxDevice *dxDev) {
             hr = d3dDev->CreateBuffer(&bufDesc, &initData, this->vertexBuf.ReleaseAndGetAddressOf());
             H::System::ThrowIfFailed(hr);
 
-            bufDesc.ByteWidth = indices.size() * sizeof(uint16_t);
+            bufDesc.ByteWidth = static_cast<UINT>(indices.size() * sizeof(uint16_t));
             bufDesc.BindFlags = D3D11_BIND_INDEX_BUFFER;
 
             initData.pSysMem = indices.data();
